requirements.txt
about-time==4.2.1
alive-progress==3.2.0
colorama==0.4.6
coverage==5.5
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
grapheme==0.6.0
libindic-soundex==1.0.2
libindic-utils==1.0.3
mutagen==1.47.0
packaging==24.2
plyer==2.1.0
Send2Trash==1.8.3
tinytag==2.0.0
toml==0.10.2
tqdm==4.67.1
